Change Directory Behavior
If the `--directory` flag is included, it is assumed you want to
add the files relative to that even if the files are passed as
absolute paths. This allows you to add the files using tab
completion from the command line.

If the flag is ommited, then the file names are dependent on if
if is a relative or absolute path. All paths are cleaned up with
any `.` and `..` directory names resolved.